% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_models.R
\name{geom_lm_line}
\alias{geom_lm_line}
\title{geom_lm_line}
\source{
https://gist.github.com/shenqi/29b42a6be83ec9e0517d
}
\usage{
geom_lm_line(model, x = NULL, ...)
}
\arguments{
\item{model}{(lm object) the linear model object to use}

\item{x}{????}

\item{...}{other arguments to be passed on to geom_abline()}
}
\description{
Convenience function for ggplot to play friendly with lm models
}
\author{
Shenqi
}

