% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_models.R
\name{ggplot_model}
\alias{ggplot_model}
\title{ggplot_model}
\usage{
ggplot_model(model, points = TRUE, residuals = FALSE, model.color = "red",
  model.width = 1, model.alpha = 1, point.col = "blue",
  point.alpha = 0.4, point.size = 6, point.border = 1,
  resid.col = "orange", resid.alpha = 0.7, resid.width = 1, ...)
}
\arguments{
\item{model}{(model object) an object created from a modelling function such
as lm()}

\item{points}{(logical) Should it show the observed data points?

     (DEFAULT = TRUE)}

\item{residuals}{(logical) Should it show the residuals as vertical lines?

     (DEFAULT = FALSE)}

\item{model.color}{(string) The color of the model line/curve

     (DEFAULT = "red")}

\item{model.width}{(numeric) width of the model line/curve

      (DEFAULT = 1)}

\item{model.alpha}{(numeric) alpha of the model line/curve

      (DEFAULT = 1)}

\item{point.col}{(string) The color for the points

     (DEFAULT = "blue")}

\item{point.size}{(int) The size of the points

     (DEFAULT = 4)}

\item{point.border}{(int) size of border of points

     (DEFAULT = 1)}

\item{resid.col}{(string) the color for the residual lines

     (DEFAULT = "red")}

\item{resid.alpha}{(numeric) alpha for the residual lines

     (DEFAULT = 0.7)}

\item{resid.width}{(numeric) width of the residual lines

     (DEFAULT = 1)}

\item{...}{aditional parameters for the ggplot}
}
\description{
Create a plot of some particular model object. You can chose to plot the
observed values and the residuals as well.
}
\details{
Currently only supports linear models, but support will be expanded to any
arbitrary model in the future.
}
\examples{
data(iris)
model = lm(iris$Petal.Length ~ iris$Sepal.Length)
ggplot_model(model, points=TRUE, residuals=TRUE, point.col="green")

ggplot_model(model, points=FALSE, residuals=TRUE, point.size=4, resid.width=5, resid.alpha=0.2, resid.col="red", point.border=2)
}
\author{
Ronny Restrepo
}

