% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggplot_models.R
\name{ggplot_residuals}
\alias{ggplot_residuals}
\title{ggplot_residuals}
\usage{
ggplot_residuals(model, points = TRUE, lines = FALSE, point.col = "blue",
  point.alpha = 0.4, point.size = 6, point.border = 1,
  line.col = "orange", line.alpha = 0.7, line.width = 1, ...)
}
\arguments{
\item{model}{(model object) an object created from a modelling function such
as lm()}

\item{points}{(logical) Should it show the residuals as points?

     (DEFAULT = TRUE)}

\item{lines}{(logical) Should it show the residuals as vertical lines?

     (DEFAULT = FALSE)}

\item{point.col}{(string) The color for the points

     (DEFAULT = "blue")}

\item{point.alpha}{(numeric) The alpha value for the points

     (DEFAULT = 0.4)}

\item{point.size}{(numeric) The size of the points

     (DEFAULT = 6)}

\item{point.border}{(numeric) size of border of points

     (DEFAULT = 1)}

\item{line.col}{(string) the color for the lines

     (DEFAULT = "orange")}

\item{line.alpha}{(numeric) alpha value for the lines

     (DEFAULT = 0.7)}

\item{line.width}{(numeric) the width of the lines

     (DEFAULT = 1)}

\item{...}{aditional parameters for the ggplot}
}
\description{
Create a plot of observed x values vs residuals for some particular model
object.
}
\details{
You can chose to either display the residuals as points, or as vertical line
segments, or both by setting the arguments "points", and "lines" to TRUE or
FALSE.
}
\examples{
data(iris)
model = lm(iris$Sepal.Width ~ iris$Petal.Width)
ggplot_residuals(model, lines=TRUE)

ggplot_residuals(model, points=FALSE, lines=TRUE, line.width=7,
line.alpha=0.2, line.col="red")
}
\author{
Ronny Restrepo
}

